[NEAR-126] Chore: 모니터링 라우터 어드민 라이브러리 이용
